requirements.txt
click==8.1.8
einops==0.8.2
exceptiongroup==1.3.1
filelock==3.19.1
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.8.0
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python-headless==4.7.0.72
packaging==26.0
pandas==2.1.3
pillow==11.3.0
Pygments==2.20.0
PyYAML==6.0.3
rich==14.3.3
safetensors==0.7.0
scikit-image==0.22.0
shellingham==1.5.4
sympy==1.14.0
timm==1.0.26
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.3
typer==0.23.2
typing_extensions==4.15.0
